+++ spimdisasm/mips/sections/MipsSectionRelocZ64.py
@@ -98,6 +98,7 @@ class SectionRelocZ64(SectionBase):
         sym = symbols.SymbolData(self.context, vrom, vromEnd, localOffset + self.inFileOffset, currentVram, self.words[0:4], self.segmentVromStart, self.overlayCategory)
         sym.contextSym.name = f"{self.name}_OverlayInfo"
         sym.contextSym.userDeclaredType = "s32"
+        sym.contextSym.allowedToReferenceSymbols = False
         sym.parent = self
         sym.setCommentOffset(self.commentOffset)
         sym.endOfLineComment = {i: f" /* _{self.name}Segment{sectName.toCapitalizedStr()}Size */" for i, sectName in enumerate(common.FileSections_ListBasic)}
@@ -111,6 +112,7 @@ class SectionRelocZ64(SectionBase):
         sym = symbols.SymbolData(self.context, vrom, vromEnd, localOffset + self.inFileOffset, currentVram, [self.relocCount], self.segmentVromStart, self.overlayCategory)
         sym.contextSym.name = f"{self.name}_RelocCount"
         sym.contextSym.userDeclaredType = "s32"
+        sym.contextSym.allowedToReferenceSymbols = False
         sym.parent = self
         sym.setCommentOffset(self.commentOffset)
         sym.analyze()
@@ -123,6 +125,7 @@ class SectionRelocZ64(SectionBase):
         sym = symbols.SymbolData(self.context, vrom, vromEnd, localOffset + self.inFileOffset, currentVram, [r.reloc for r in self.entries], self.segmentVromStart, self.overlayCategory)
         sym.contextSym.name = f"{self.name}_OverlayRelocations"
         sym.contextSym.userDeclaredType = "s32"
+        sym.contextSym.allowedToReferenceSymbols = False
         sym.parent = self
         sym.setCommentOffset(self.commentOffset)
         sym.endOfLineComment = {i: f" /* {str(r)} */" for i, r in enumerate(self.entries)}
@@ -135,8 +138,9 @@ class SectionRelocZ64(SectionBase):
             vrom = self.getVromOffset(localOffset)
             vromEnd = vrom + 4 * len(self.tail)
             sym = symbols.SymbolData(self.context, vrom, vromEnd, localOffset + self.inFileOffset, currentVram, self.tail, self.segmentVromStart, self.overlayCategory)
-            sym.contextSym.userDeclaredType = "s32"
             sym.contextSym.name = f"{self.name}_Padding"
+            sym.contextSym.userDeclaredType = "s32"
+            sym.contextSym.allowedToReferenceSymbols = False
             sym.parent = self
             sym.setCommentOffset(self.commentOffset)
             sym.analyze()
@@ -149,6 +153,7 @@ class SectionRelocZ64(SectionBase):
         sym = symbols.SymbolData(self.context, vrom, vromEnd, localOffset + self.inFileOffset, currentVram, [self.seekup], self.segmentVromStart, self.overlayCategory)
         sym.contextSym.name = f"{self.name}_OverlayInfoOffset"
         sym.contextSym.userDeclaredType = "s32"
+        sym.contextSym.allowedToReferenceSymbols = False
         sym.parent = self
         sym.setCommentOffset(self.commentOffset)
         sym.analyze()